refactor home page to use server loaded tab config
